work on p76
I've hit a bit of a snag with p84. I know how I want to compute it: Markov Chains, but it definitely will take some time to build the probability matrix procedurally.

As for p76, I have some pretty cool ideas which hinge on a recursive formula for the number of combinations one could use to arrive at a desired number n. Much of the formula is found in p76, the trick is really getting my alpha(a,n) function working.